Rearrange task definition class
